AI-interview-chatbot-main/requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0

# Authentication & Authorization
google-auth-oauthlib==1.0.0
google-auth==2.22.0
google-api-python-client==2.95.0

# Calendar Integration
O365==2.0.26
icalendar==5.0.7
recurring-ical-events==2.0.2
pytz==2023.3

# ML & Analytics
mlflow==2.8.0
wandb==0.15.5
optuna==3.2.0

# Deep Learning & Computer Vision
torch==2.1.0
torchvision==0.16.0
transformers==4.34.0
huggingface-hub>=0.16.4,<0.17.0
tokenizers>=0.14.0,<0.15.0
opencv-python==4.8.0.74
opencv-contrib-python==4.8.0.74

# LLM & Vector Search
langchain==0.0.200
faiss-cpu==1.7.4
openai==0.27.8

# Visualization
plotly==5.15.0
seaborn==0.12.2
matplotlib==3.7.1
kaleido==0.2.1

# Audio Processing
librosa==0.10.0
soundfile==0.12.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1
aiohttp==3.8.5
python-dateutil==2.8.2

# Document Processing
openpyxl==3.1.2
xlrd==2.0.1
xlsxwriter==3.1.0
pdfkit==1.0.0
jinja2==3.1.2
WeasyPrint==59.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Code Quality
black==23.3.0
flake8==6.0.0
isort==5.12.0

# Streamlit Components
streamlit-webrtc>=0.62.0

# Additional dependencies
alembic==1.13.1
redis==5.0.1
websockets==12.0
